+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.1] - 2025-10-30
+
+### Fixed
+- Improved Edit mode compatibility by using different mesh data access methods
+- Fixed vertex group weight access in Edit mode using vertex.groups instead of vg.weight()
+
+## [1.2.0] - 2025-10-30
+
+### Added
+- Support for Edit mode in addition to Weight Paint mode
+- Vertex weights are now visible in both Weight Paint and Edit modes
+
+### Changed
+- Updated description to reflect new Edit mode support
+
 ## [1.1.2] - 2025-10-30
 
 ### Changed